feature:标准设置页面样式
